Refs #10364 email addresses visible, show user id instead of name
To use the name would require also changing all the translations,
which is probably a bit excessive for the sake of the deletion
log. This way we can still easily link through to the user
contributions page. Even if the user has since changed ID, a visitor
will be redirected.